Add files via upload
The Artist Lab picture says that I did't complete the entirety of the ten steps, however the page crashed three and reset all my progress. I completed the exercise twice because of this and so on the third try, I just jumped to the 10th step, since I already did it twice. If you'd like me too do the exercise again, please let me know and I'll comply